:truck: Moved overview components out to now dir
