+++ aksel.nav.no/website/app/_sanity/queries.ts
@@ -64,7 +64,7 @@ const KOMPONENT_BY_SLUG_QUERY =
     },
 }`);
 
-const KOMPONENTOVERSIKT_QUERY =
+const DESIGNSYSTEM_OVERVIEW_BY_CATEGORY_QUERY =
   defineQuery(`*[_type == "komponent_artikkel"  && kategori == $category]
   {
     _id,
@@ -131,7 +131,7 @@ export {
   DESIGNSYSTEM_KOMPONENTER_LANDINGPAGE_QUERY,
   GLOBAL_SEARCH_QUERY_ALL,
   KOMPONENT_BY_SLUG_QUERY,
-  KOMPONENTOVERSIKT_QUERY,
+  DESIGNSYSTEM_OVERVIEW_BY_CATEGORY_QUERY,
   TOC_BY_SLUG_QUERY,
   SLUG_BY_TYPE_QUERY,
   GRUNNLEGGENDE_BY_SLUG_QUERY,
